+++ tests/unit/clients/test_jinja_static.py
