+++ tests/unit/clients/test_registry.py
